Add replica oracle to Postgress OrdenPago

+++ Convertidor/Data/EntitiesDestino/ADM/ADM_RETENCIONES_OP.cs
@@ -6,8 +6,8 @@
         public int CODIGO_ORDEN_PAGO { get; set; }
         public int TIPO_RETENCION_ID { get; set; }
         public int? CODIGO_RETENCION { get; set; }
-        public int? POR_RETENCION { get; set; }
-        public int? MONTO_RETENCION { get; set; }
+        public decimal? POR_RETENCION { get; set; }
+        public decimal? MONTO_RETENCION { get; set; }
         public string? EXTRA1 { get; set; }=string.Empty;
         public string? EXTRA2 { get; set; }= string.Empty;
         public string? EXTRA3 { get; set; } = string.Empty;
@@ -18,7 +18,7 @@
         public int CODIGO_EMPRESA { get; set; }
         public int? CODIGO_PRESUPUESTO { get; set; }
         public string? EXTRA4 { get; set; } = string.Empty;
-        public int? BASE_IMPONIBLE { get; set; }
+        public decimal? BASE_IMPONIBLE { get; set; }
 
     }
 }
